+++ src/Oro/Bundle/FilterBundle/Expression/Date/Compiler.php
@@ -2,6 +2,9 @@
 
 namespace Oro\Bundle\FilterBundle\Expression\Date;
 
+/**
+ * Responsible for compiling the datetime.
+ */
 class Compiler
 {
     /** @var Lexer */
@@ -24,5 +27,17 @@ class Compiler
     public function compile($string, $returnRawToken = false)
     {
         return $this->parser->parse($this->lexer->tokenize($string), $returnRawToken);
+    }
+
+    /**
+     * @param string $date
+     * @param bool $returnRawToken
+     * @param string|null $timeZone
+     *
+     * @return int|mixed|ExpressionResult|string
+     */
+    public function compileWithTimeZone(string $date, bool $returnRawToken = false, ?string $timeZone = null)
+    {
+        return $this->parser->parseWithTimeZone($this->lexer->tokenize($date), $returnRawToken, $timeZone);
     }
 }