+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Please refer first to [UPGRADE.md](UPGRADE.md) for the most important items that should be addressed before attempting to upgrade or during the upgrade of a vanilla Oro application.
 
 The current file describes significant changes in the code that may affect the upgrade of your customizations.
+
+## 4.1.10
+### Changed
+
+#### SSOBundle
+* The configuration option `oro_sso.enable_google_sso` was renamed to `oro_google_integration.enable_sso`.
+* The configuration option `oro_sso.domains` was renamed to `oro_google_integration.sso_domains`.
+* The service `oro_sso.oauth_provider` was renamed to `oro_sso.oauth_user_provider`.
+
 ## 4.1.7
 ### Changed